# docs/openapi-backup.yaml
...
openapi: 3.1.0
...
info:
  # Required fields
  title: trAIner API
  version: 1.0.0
  
  # Optional but recommended fields
  description: |
    The trAIner API provides comprehensive endpoints for AI-powered fitness tracking and personalized workout planning. 
    
    ## Overview
    This API enables:
    - User authentication and profile management via Supabase
    - AI-driven workout plan generation and customization
    - Progress tracking and analytics with AI insights
    - Nutrition logging and macro calculations
    - Data export/import functionality
    - Real-time health monitoring
    
    ## Authentication
    All protected endpoints require JWT authentication via Supabase Auth.
    
    ## Rate Limiting
    API requests are rate-limited to ensure service quality:
    - Authentication endpoints: 10 requests/15 minutes
    - Workout generation: 10 requests/hour
    - Data export: 3 requests/hour
    - Standard operations: 1000 requests/hour
    
    ## Data Security
    - All data transmission uses HTTPS encryption
    - JWT tokens for secure session management
    - Input validation and SQL injection protection
    - Medical data handled with appropriate security measures
  
  # Contact information (to be updated before deployment)
  contact:
    name: trAIner API Support
    url: https://trainer.app/support
    email: support@trainer.app
  
  # Terms of service URL (to be created before deployment)
  termsOfService: https://trainer.app/terms

...
servers:
...

...
components:
  securitySchemes:
    SupabaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Supabase Auth. Include in Authorization header as 'Bearer {token}'.
        Tokens expire based on Supabase configuration and must be refreshed using the refresh token.

  # Authentication & User Management Schemas
  schemas:
...
    WorkoutGenerationRequest:
      type: object
      required: [fitnessLevel, goals]
      properties:
        fitnessLevel:
          type: string
          enum: [beginner, intermediate, advanced]
          description: User's fitness experience level
          example: intermediate
        goals:
          type: array
          items:
            type: string
            enum: [weight_loss, muscle_gain, strength, endurance, flexibility, general_fitness]
          minItems: 1
          maxItems: 5
          description: User's fitness objectives (1-5 goals)
          example: ["muscle_gain", "strength"]
        equipment:
          type: array
          items:
            type: string
          nullable: true
          description: Available fitness equipment
          example: ["dumbbells", "barbell", "resistance_bands"]
        restrictions:
          type: array
          items:
            type: string
          nullable: true
          description: Physical limitations or injuries to avoid
          example: ["lower_back_pain", "knee_injury"]
        exerciseTypes:
          type: array
          items:
            type: string
            enum: [cardio, strength, hiit, yoga, flexibility, sports_specific]
          nullable: true
          description: Preferred exercise categories
          example: ["strength", "cardio"]
        workoutFrequency:
          type: string
          nullable: true
          description: Preferred workout schedule
          example: "3x per week"
        targetDuration:
          type: integer
          minimum: 15
          maximum: 180
          nullable: true
          description: Target workout duration in minutes
          example: 60
        experienceDetails:
          type: object
          nullable: true
          properties:
            yearsTraining:
              type: integer
              minimum: 0
              maximum: 50
              description: Years of training experience
            specificSports:
              type: array
              items:
                type: string
              description: Sports or activities with experience
            currentRoutine:
              type: string
              description: Description of current workout routine
          description: Additional experience context for personalization
      example:
        fitnessLevel: intermediate
        goals: ["muscle_gain", "strength"]
        equipment: ["dumbbells", "barbell"]
        restrictions: ["lower_back_pain"]
        exerciseTypes: ["strength"]
        workoutFrequency: "4x per week"
        targetDuration: 75
...
    GenericError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [error]
          description: Response status indicator
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details (development only)
      example:
        status: error
        message: An unexpected error occurred
...
  responses:
...
    ValidationFailed:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'

    # Workout Management Success Responses
    WorkoutPlanGenerated:
      description: Workout plan generated successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutPlanResponse'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
    WorkoutPlanRetrieved:
      description: Workout plan retrieved successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutPlanResponse'
    WorkoutPlansListed:
      description: Workout plans list retrieved successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutPlansListResponse'
    WorkoutPlanAdjusted:
      description: Workout plan adjusted successfully
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutAdjustmentResponse'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
...
    WorkoutValidationFailed:
      description: Workout request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutValidationError'
    WorkoutGenerationFailed:
      description: AI workout generation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutGenerationError'
    WorkoutNotFound:
      description: Workout plan not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutNotFoundError'
    WorkoutPermissionDenied:
      description: Permission denied for workout operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutPermissionError'
    WorkoutAdjustmentFailed:
      description: Workout adjustment failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutAdjustmentError'
    WorkoutRateLimitExceeded:
      description: Workout operation rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/WorkoutRateLimitError'
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'

  # Common Response Headers
  headers:
    X-RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        type: integer
    X-RateLimit-Remaining:
      description: The number of remaining requests in the current period
      schema:
        type: integer
    X-RateLimit-Reset:
      description: The time at which the current rate limit window resets (Unix timestamp)
      schema:
        type: integer
...
security:
  - SupabaseJWT: []

# 6. Tags Definition (Based on Verified Route Structure)
...
paths:
...
  /workouts:
    post:
      tags:
        - Workout Management
      summary: Generate AI-powered workout plan
      description: |
        Generate a personalized workout plan using AI agents. The system combines user profile data,
        exercise research from multiple sources, and AI-powered plan generation to create customized
        workout plans. Rate limited to 10 requests per hour per user.
      operationId: generateWorkoutPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutGenerationRequest'
            examples:
              beginner_strength:
                summary: Beginner strength training
                value:
                  fitnessLevel: beginner
                  goals: ["muscle_gain", "strength"]
                  equipment: ["dumbbells", "resistance_bands"]
                  exerciseTypes: ["strength"]
                  workoutFrequency: "3x per week"
                  targetDuration: 45
              intermediate_mixed:
                summary: Intermediate mixed training
                value:
                  fitnessLevel: intermediate
                  goals: ["endurance", "weight_loss"]
                  equipment: ["bodyweight", "cardio_machines"]
                  exerciseTypes: ["cardio", "strength"]
                  restrictions: ["knee_injury"]
                  workoutFrequency: "4x per week"
                  targetDuration: 60
      responses:
        '201':
          $ref: '#/components/responses/WorkoutPlanGenerated'
        '400':
          $ref: '#/components/responses/WorkoutValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/WorkoutRateLimitExceeded'
        '500':
          $ref: '#/components/responses/WorkoutGenerationFailed'

    get:
      tags:
        - Workout Management
      summary: List user's workout plans
      description: |
        Retrieve a paginated list of workout plans for the authenticated user.
        Supports filtering by status, difficulty, and search terms. Plans are sorted
        by creation date (newest first) by default.
      operationId: listWorkoutPlans
      parameters:
        - name: limit
          in: query
          description: Maximum number of plans to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of plans to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: searchTerm
          in: query
          description: Search term for plan names or descriptions
          required: false
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: status
          in: query
          description: Filter by plan status
          required: false
          schema:
            type: string
            enum: [draft, active, archived]
        - name: difficulty
          in: query
          description: Filter by difficulty level
          required: false
          schema:
            type: string
            enum: [beginner, intermediate, advanced]
        - name: sortBy
          in: query
          description: Field to sort results by
          required: false
          schema:
            type: string
            enum: [created_at, updated_at, name]
            default: created_at
        - name: sortOrder
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          $ref: '#/components/responses/WorkoutPlansListed'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workouts/{planId}:
    get:
      tags:
        - Workout Management
      summary: Get specific workout plan
      description: |
        Retrieve detailed information about a specific workout plan including all exercises,
        AI reasoning, and generation metadata. Only the plan owner can access the plan.
      operationId: getWorkoutPlan
      parameters:
        - name: planId
          in: path
          description: Unique workout plan identifier
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          $ref: '#/components/responses/WorkoutPlanRetrieved'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/WorkoutPermissionDenied'
        '404':
          $ref: '#/components/responses/WorkoutNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Workout Management
      summary: Adjust workout plan with AI
      description: |
        Modify an existing workout plan using AI-powered adjustment. Provide natural language
        feedback or structured adjustment instructions. The AI will analyze the feedback and
        apply appropriate modifications while maintaining safety and effectiveness.
        Rate limited to 20 requests per hour per user.
      operationId: adjustWorkoutPlan
      parameters:
        - name: planId
          in: path
          description: Unique workout plan identifier
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkoutAdjustmentRequest'
            examples:
              natural_language:
                summary: Natural language adjustment
                value:
                  adjustments: "Make the workout more challenging and add more compound movements"
                  userContext:
                    recentPerformance: "Workouts have been feeling too easy"
                    energyLevels: high
              structured_adjustment:
                summary: Structured adjustment request
                value:
                  adjustments:
                    exercisesToAdd: ["deadlifts", "pull-ups"]
                    exercisesToRemove: ["leg extensions"]
                    notesOrPreferences: "Focus more on functional movements"
                    intensityAdjustment: increase
      responses:
        '200':
          $ref: '#/components/responses/WorkoutPlanAdjusted'
        '400':
          $ref: '#/components/responses/WorkoutValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/WorkoutPermissionDenied'
        '404':
          $ref: '#/components/responses/WorkoutNotFound'
        '422':
          $ref: '#/components/responses/WorkoutAdjustmentFailed'
        '429':
          $ref: '#/components/responses/WorkoutRateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Workout Management
      summary: Delete workout plan
      description: |
        Permanently delete a workout plan. This action cannot be undone. The plan
        will be soft deleted to maintain data integrity for any associated workout logs.
        Only the plan owner can delete the plan.
      operationId: deleteWorkoutPlan
      parameters:
        - name: planId
          in: path
          description: Unique workout plan identifier
          required: true
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '204':
          description: Workout plan deleted successfully (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/WorkoutPermissionDenied'
        '404':
          $ref: '#/components/responses/WorkoutNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# Additional OpenAPI elements will be added in subsequent phases:
# - Additional paths for other categories (Phase 3)
# - More component schemas (Phase 3)
# - Component responses (Phase 4)
# - Component parameters (Phase 5)
# - Component examples (Phase 6)

# Additional root-level extensions for API governance
...